Reorganise .gitignore files
The generated CSS file (generated by compiling style.less) goes in
the css directory, so it’s clearer to organise the .gitignore files
in such a way that the directory is created when the repository is
cloned.

Also add a logs directory, for putting the Apache logs in when the
server is used locally for development.